Quantiation Parameters

Quantitation of mass spectrometry data is implemented through the quantify() method, that converts raw data objects such as Spectra() into QFeatures() objects that store and hande quantitative data. Quantitation methods for MS data a numerous and varied, and the exact method to be applied on the raw data is paramterised by a QuantParam instance, i.e. a dedicated class that stores all the parameters needed to perform a specific quantitation method.

MS quantitation can be performed at the MS1 or MS2 (and 3) levels. Additionally, quantitation can be label-free or labelled. These two states are defined by the msLevel (an integer) and label (a logical) parameters.

Additional details about the quantitation method are defined as a named list, as described in the 'Details' section. The example below illustrates how to defined labelled (label = TRUE) MS2 (msLevel = 2L) quantitation using TMT 10-plex isobaric tagging (defined by the TMT10` ReporterIons() instance).

# S4 method for QuantParam
isEmpty(x)

QuantParam(msLevel = NA_integer_, label = NA, params = list())

Arguments

x
A instance of class QuantParam.

msLevel
integer(1) indicating the MS levels to be quantified. Default is NA.

label
logical(1) defining is labelled or label-free quantitation. Default is NA.

params
list() containing additional method-specific quantitation parameters. The list must be named.

Details

Labelled MS2 Quantitation:

The params need only to contain a single mandatory element, a ReporterIons() instance, named reporters.

Labelled MS3 Quantitation:

The params need to contain a mandatory element, a ReporterIons() instance, named reporters.

Two optional character strings, named acquisitionNum and precScanNum, containing the respective column names of the acquition number and precursor scan acquisition numbers in the raw data to be quantified. The default values are "acquisitionNum" and "precScanNum" respectively, which match the names in standard Spectra objects.

Slots

msLevel

integer(1) indicating the MS level the quantitation will be performed on.

label

params

named list() of additional parameters.

Author

Laurent Gatto

Examples

## default (empty) parameters
QuantParam()

#> Object of class:  QuantParam 
#>  Parameters:
#>  - msLevel: [1] NA
#>  - label: [1] NA
#>  - params: list()

isEmpty(QuantParam())

#> [1] TRUE

## MS2 quantitation using TMT10 plex
QuantParam(msLevel = 2L,
           label = TRUE,
           params = list(reporters = TMT10))

#> Object of class:  QuantParam 
#>  Parameters:
#>  - msLevel: [1] 2
#>  - label: [1] TRUE
#>  - params: $reporters
#> Object of class "ReporterIons"
#> TMT10HCD' with 10 reporter ions
#>  - [TMT10_126] 126.1277 +/- 0.002
#>  - [TMT10_127N] 127.1248 +/- 0.002
#>  - [TMT10_127C] 127.1311 +/- 0.002
#>  - [TMT10_128N] 128.1281 +/- 0.002
#>  - [TMT10_128C] 128.1344 +/- 0.002
#>  - [TMT10_129N] 129.1315 +/- 0.002
#>  - [TMT10_129C] 129.1378 +/- 0.002
#>  - [TMT10_130N] 130.1348 +/- 0.002
#>  - [TMT10_130C] 130.1411 +/- 0.002
#>  - [TMT10_131] 131.1382 +/- 0.002
#> 

## MS3 quantitation using TMT16 plex and non-standard acquisition
## number and precursor scan acqusition number
QuantParam(msLevel = 2L, label = TRUE,
           params = list(reporters = TMT10,
                         acqusitionNum = "aquisition_number",
                         preScanNum = "prec_scan_number"))

#> Object of class:  QuantParam 
#>  Parameters:
#>  - msLevel: [1] 2
#>  - label: [1] TRUE
#>  - params: $reporters
#> Object of class "ReporterIons"
#> TMT10HCD' with 10 reporter ions
#>  - [TMT10_126] 126.1277 +/- 0.002
#>  - [TMT10_127N] 127.1248 +/- 0.002
#>  - [TMT10_127C] 127.1311 +/- 0.002
#>  - [TMT10_128N] 128.1281 +/- 0.002
#>  - [TMT10_128C] 128.1344 +/- 0.002
#>  - [TMT10_129N] 129.1315 +/- 0.002
#>  - [TMT10_129C] 129.1378 +/- 0.002
#>  - [TMT10_130N] 130.1348 +/- 0.002
#>  - [TMT10_130C] 130.1411 +/- 0.002
#>  - [TMT10_131] 131.1382 +/- 0.002
#> 
#> $acqusitionNum
#> [1] "aquisition_number"
#> 
#> $preScanNum
#> [1] "prec_scan_number"
#>