Reporter Ions Class

The ReporterIons class defines a set of isobaric reporter ions that are used for quantification of labelled MS2 data, e.g. iTRAQ (isobaric tag for relative and absolute quantitation) or TMT (tandem mass tags).

Many commercial reporter ions are readily available. Custom instances can be created with the ReporterIons() function.

ReporterIons(name, reporterNames, mz, width)

# S4 method for ReporterIons
show(object)

# S4 method for ReporterIons,ANY,ANY,ANY
[(x, i = "numeric", j = "missing", drop = "missing")

# S4 method for ReporterIons
length(x)

# S4 method for ReporterIons
mz(object)

# S4 method for ReporterIons
width(x)

reporterNames(x)

reporterNames(x) <- value

# S4 method for ReporterIons
names(x)

iTRAQ4

iTRAQ5

iTRAQ8

iTRAQ9

TMT10HCD

TMT10

TMT10ETD

TMT11HCD

TMT11

TMT16HCD

TMT16

TMT6

TMT6b

TMT7

TMT7b

Arguments

name
Parameter to set the name slot.

reporterNames
Parameter to set the reporterNames slot.

mz
Parameter to set the mz slot.

width
Parameter to set the width slot.

object
An instance of class ReporterIons.

x
An instance of class ReporterIons.

i
numeric() to subset x.

j
ignored.

drop
ignored.

value
A value for replacement.

Format

An object of class ReporterIons of length 4.

An object of class ReporterIons of length 5.

An object of class ReporterIons of length 8.

An object of class ReporterIons of length 9.

An object of class ReporterIons of length 10.

An object of class ReporterIons of length 10.

An object of class ReporterIons of length 10.

An object of class ReporterIons of length 11.

An object of class ReporterIons of length 11.

An object of class ReporterIons of length 16.

An object of class ReporterIons of length 16.

An object of class ReporterIons of length 6.

An object of class ReporterIons of length 6.

An object of class ReporterIons of length 7.

An object of class ReporterIons of length 7.

Slots

name

character(1) naming the instance.

reporterNames

character() of length equal to the number of reporter ions in the instance. Used to uniquely name each ion.

mz

numeric() with the reporter ions' m/z values.

width

numeric(1) defining the region where the reporter ion can be expected. This region is calculated as mz +/- width for earch reporter ion.

References

Ross PL, Huang YN, Marchese JN, Williamson B, Parker K, Hattan S, Khainovski N, Pillai S, Dey S, Daniels S, Purkayastha S, Juhasz P, Martin S, Bartlet-Jones M, He F, Jacobson A, Pappin DJ. Multiplexed protein quantitation in Saccharomyces cerevisiae using amine-reactive isobaric tagging reagents. Mol Cell Proteomics, 2004 Dec;3(12):1154-69. Epub 2004 Sep 22. PubMed PMID: 15385600.

Thompson A, Schäfer J, Kuhn K, Kienle S, Schwarz J, Schmidt G, Neumann T, Johnstone R, Mohammed AK, Hamon C. Tandem mass tags: a novel quantification strategy for comparative analysis of complex protein mixtures by MS/MS. Anal Chem. 2003 Apr 15;75(8):1895-904. Erratum in: Anal Chem. 2006 Jun 15;78(12):4235. Mohammed, A Karim A and Anal Chem. 2003 Sep 15;75(18):4942. Johnstone, R. PubMed PMID: 12713048.

Author

Laurent Gatto

Examples

## Manual construction of a ReporterIons instance:
ReporterIons(name = "iTRAQ4",
             reporterNames = c("iTRAQ4.114", "iTRAQ4.115",
                               "iTRAQ4.116", "iTRAQ4.117"),
             mz = c(114.1, 115.1, 116.1, 117.1),
             width = 0.05)

#> Object of class "ReporterIons"
#> iTRAQ4' with 4 reporter ions
#>  - [iTRAQ4.114] 114.1 +/- 0.05
#>  - [iTRAQ4.115] 115.1 +/- 0.05
#>  - [iTRAQ4.116] 116.1 +/- 0.05
#>  - [iTRAQ4.117] 117.1 +/- 0.05

## Some pre-defined reporter ions constructors:
TMT10

#> Object of class "ReporterIons"
#> TMT10HCD' with 10 reporter ions
#>  - [TMT10_126] 126.1277 +/- 0.002
#>  - [TMT10_127N] 127.1248 +/- 0.002
#>  - [TMT10_127C] 127.1311 +/- 0.002
#>  - [TMT10_128N] 128.1281 +/- 0.002
#>  - [TMT10_128C] 128.1344 +/- 0.002
#>  - [TMT10_129N] 129.1315 +/- 0.002
#>  - [TMT10_129C] 129.1378 +/- 0.002
#>  - [TMT10_130N] 130.1348 +/- 0.002
#>  - [TMT10_130C] 130.1411 +/- 0.002
#>  - [TMT10_131] 131.1382 +/- 0.002

iTRAQ4

#> Object of class "ReporterIons"
#> iTRAQ4' with 4 reporter ions
#>  - [iTRAQ4_114] 114.1112 +/- 0.05
#>  - [iTRAQ4_115] 115.1083 +/- 0.05
#>  - [iTRAQ4_116] 116.1116 +/- 0.05
#>  - [iTRAQ4_117] 117.115 +/- 0.05

TMT10[1:5]

#> Object of class "ReporterIons"
#> TMT10HCD[1:2:3:4:5]' with 5 reporter ions
#>  - [TMT10_126] 126.1277 +/- 0.002
#>  - [TMT10_127N] 127.1248 +/- 0.002
#>  - [TMT10_127C] 127.1311 +/- 0.002
#>  - [TMT10_128N] 128.1281 +/- 0.002
#>  - [TMT10_128C] 128.1344 +/- 0.002

reporterNames(TMT10)

#>  [1] "TMT10_126"  "TMT10_127N" "TMT10_127C" "TMT10_128N" "TMT10_128C"
#>  [6] "TMT10_129N" "TMT10_129C" "TMT10_130N" "TMT10_130C" "TMT10_131" 

names(TMT10)

#> [1] "TMT10HCD"

mz(TMT10)

#>  [1] 126.1277 127.1248 127.1311 128.1281 128.1344 129.1315 129.1378 130.1348
#>  [9] 130.1411 131.1382

width(TMT10)

#> [1] 0.002

as(TMT10, "DataFrame")

#> DataFrame with 10 rows and 4 columns
#>           name reporterName        mz     width
#>    <character>  <character> <numeric> <numeric>
#> 1     TMT10HCD    TMT10_126   126.128     0.002
#> 2     TMT10HCD   TMT10_127N   127.125     0.002
#> 3     TMT10HCD   TMT10_127C   127.131     0.002
#> 4     TMT10HCD   TMT10_128N   128.128     0.002
#> 5     TMT10HCD   TMT10_128C   128.134     0.002
#> 6     TMT10HCD   TMT10_129N   129.131     0.002
#> 7     TMT10HCD   TMT10_129C   129.138     0.002
#> 8     TMT10HCD   TMT10_130N   130.135     0.002
#> 9     TMT10HCD   TMT10_130C   130.141     0.002
#> 10    TMT10HCD    TMT10_131   131.138     0.002